Fix existing test instead of adding a separate one
Configure deployment options (without useVersioning) on the existing
ActivityCancellationTokenIntegrationTest worker, register a nexus
service so the normal poller starts, and assert both sides: normal
polls carry deploymentOptions, worker-command polls do not.
